+++ Example/Pods/Target Support Files/KeyboardAvoidingView/KeyboardAvoidingView-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_KeyboardAvoidingView : NSObject
-@end
-@implementation PodsDummy_KeyboardAvoidingView
-@end