+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 All notable changes to this project will be documented in this file.
 `KeyboardAvoidingView` adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.0](https://github.com/APUtils/KeyboardAvoidingView/releases/tag/2.1.0)
+Released on 2018-02-15.
+
+#### Added
+- Safe area and layout guides support
+
+#### Fixed
+- Do not animate transitions if there is no window.
+- Layout on view controller level instead of window level to prevent some animation glitches.
+
+
 ## [2.0.0](https://github.com/APUtils/KeyboardAvoidingView/releases/tag/2.0.0)
 Released on 2017-09-21.